Displays, but has no data...
The data that I need is non-existant, so when it displays, it displays
nothing.  I am thinking that I need to put something more accurate for
the display to know for sure that it is going to work.

Right now I want to look at how to generate the primes, though I'm going
to add two values to the database.