Database has data, tables updated, errors corrected
I added the first two data points for the database tables, these should
be enough to be able to generate the data that I will need along the
way.  I will probably need to test this to be certain of it.

The tables have been updated, the "unknown" table (which I keep calling
different things) has a "first" "latest" and "last" prime values.  The
First and Last are the first and last values which are factors, and the
"latest" is the prime which last was tested against the value.  These
will be stored as "prime numbers" not prime values.

Once I started to be able to work with data, a number of errors were
found, and we have corrected them as best we can tell so far.