Fixed errors with "wrong" $primenumber values
If this gets called with the wrong value for $primenumber it was giving
an error which would not display, and tries to respond with a 502 error,
which got converted to a 404 error as it wasn't finding the correct
error message for 502.

This now seems to not be having problems any more.  It might have been
working when connecting from the local machine, but I cannot confirm
that.  The sudden errors where I had not been seeing them, and with no
explanation was confusing me.